pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "redjax", email = "jackenyon@gmail.com"},
]
dependencies = [
    "jupyter>=1.0.0",
    "ipykernel>=6.27.1",
    "polars-lts-cpu>=0.20.0",
    "pandas>=2.1.4",
    "pyarrow>=14.0.1",
    "plotly>=5.18.0",
    "dynaconf>=3.2.4",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

